can log in w/gh user
Geesh. this was a hard chunk of work.
But I now can log in via Github, and save (or find) a user
to my database.

Now I need a few other controllers to render views for them
so they can see the rest of the stuff they're expecting to see